moves registration form to utilities and updates imports
