+++ resources/assets/components/pages/VoterRegistrationDrivePage/Beta/BetaPage.js
@@ -11,10 +11,10 @@ import Placeholder from '../../../utilities/Placeholder';
 import { isDevEnvironment, query } from '../../../../helpers';
 import SiteFooter from '../../../utilities/SiteFooter/SiteFooter';
 import PrimaryButton from '../../../utilities/Button/PrimaryButton';
-import StartVoterRegistrationForm from './StartVoterRegistrationForm';
 import SiteNavigationContainer from '../../../SiteNavigation/SiteNavigationContainer';
 import ScholarshipInfoBlock from '../../../blocks/ScholarshipInfoBlock/ScholarshipInfoBlock';
 import ContentfulEntryLoader from '../../../utilities/ContentfulEntryLoader/ContentfulEntryLoader';
+import StartVoterRegistrationForm from '../../../utilities/StartVoterRegistrationForm/StartVoterRegistrationForm';
 
 const BETA_VOTER_REGISTRATION_DRIVE_PAGE_QUERY = gql`
   query BetaVoterRegistrationDrivePageQuery(
